LayoutTests/http/tests/websocket/tests/hybi/undefined-protocol-expected.txt

Test that passing undefined as protocols is equivalent to not passing any protocols

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.protocol is ""
Connected
PASS ws.protocol is ""
Closed
PASS ws.protocol is ""
PASS closeEvent.wasClean is true
Exited onclose handler
PASS ws.protocol is ""
PASS successfullyParsed is true

TEST COMPLETE
